Split library code and scripts (#5)
* Fix typo in model loading from config file

* Move source code into penet module
